ingest/requirements.txt
aiodns==3.0.0
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
aleph-alpha-client==2.16.0
anyio==3.6.2
argilla==1.3.1
async-timeout==4.0.2
attrs==22.2.0
autopep8==2.0.1
backoff==2.2.1
boto3==1.26.81
botocore==1.29.81
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
dataclasses-json==0.5.7
deeplake==3.2.11
Deprecated==1.2.13
dill==0.3.6
dnspython==2.3.0
entrypoints==0.4
et-xmlfile==1.1.0
frozenlist==1.3.3
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
hub==3.0.1
humbug==0.2.8
idna==3.4
jmespath==1.0.1
joblib==1.2.0
langchain==0.0.96
loguru==0.6.0
lxml==4.9.2
Markdown==3.4.1
marshmallow==3.19.0
marshmallow-enum==1.5.1
monotonic==1.6
multidict==6.0.4
multiprocess==0.70.14
mypy-extensions==1.0.0
nltk==3.8.1
numcodecs==0.11.0
numpy==1.23.5
openai==0.26.5
openpyxl==3.1.1
packaging==23.0
pandas==1.5.3
pathos==0.3.0
Pillow==9.4.0
pinecone-client==2.2.1
pox==0.3.2
ppft==1.7.6.6
pycares==4.3.0
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.5
PyJWT==2.6.0
python-dateutil==2.8.2
python-docx==0.8.11
python-magic==0.4.27
python-pptx==0.6.21
pytz==2022.7.1
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
rfc3986==1.5.0
s3transfer==0.6.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.46
tenacity==8.2.2
tokenizers==0.13.2
tomli==2.0.1
tqdm==4.64.1
typing-inspect==0.8.0
typing_extensions==4.5.0
unstructured==0.5.1
urllib3==1.26.14
wrapt==1.14.1
XlsxWriter==3.0.8
yarl==1.8.2
